Centralize rollup.config.js implementation.
